# Minor Type.IsPrimitive improvement